+++ Plutarch/Internal/Semigroup.hs
@@ -76,7 +76,9 @@ import Plutarch.Internal.Term (
   (#$),
  )
 import PlutusCore qualified as PLC
+import PlutusCore.Builtin (KnownBuiltinType)
 import Universe (Includes)
+import UntypedPlutusCore qualified as UPLC
 
 {- | = Laws
 
@@ -242,11 +244,15 @@ newtype PAnd (a :: S -> Type) (s :: S)
     )
     via (DeriveNewtypePlutusType (PAnd a))
 
--- | @since 1.10.0
+-- | @since 1.13.0
 deriving via
   DeriveNewtypePLiftable (PAnd a) (AsHaskell a)
   instance
-    (PLiftable a, PLC.DefaultUni `Includes` PlutusRepr a) => PLiftable (PAnd a)
+    ( PLiftable a
+    , PLC.DefaultUni `Includes` PlutusRepr a
+    , KnownBuiltinType (UPLC.Term UPLC.DeBruijn UPLC.DefaultUni UPLC.DefaultFun ()) (PlutusRepr a)
+    ) =>
+    PLiftable (PAnd a)
 
 -- | @since 1.10.0
 instance PSemigroup (PAnd PBool) where
@@ -304,11 +310,15 @@ newtype POr (a :: S -> Type) (s :: S)
     )
     via (DeriveNewtypePlutusType (POr a))
 
--- | @since 1.10.0
+-- | @since 1.13.0
 deriving via
   DeriveNewtypePLiftable (POr a) (AsHaskell a)
   instance
-    (PLiftable a, PLC.DefaultUni `Includes` PlutusRepr a) => PLiftable (POr a)
+    ( PLiftable a
+    , PLC.DefaultUni `Includes` PlutusRepr a
+    , KnownBuiltinType (UPLC.Term UPLC.DeBruijn UPLC.DefaultUni UPLC.DefaultFun ()) (PlutusRepr a)
+    ) =>
+    PLiftable (POr a)
 
 -- | @since 1.10.0
 instance PSemigroup (POr PBool) where
@@ -366,11 +376,15 @@ newtype PXor (a :: S -> Type) (s :: S)
     )
     via (DeriveNewtypePlutusType (PXor a))
 
--- | @since 1.10.0
+-- | @since 1.13.0
 deriving via
   DeriveNewtypePLiftable (PXor a) (AsHaskell a)
   instance
-    (PLiftable a, PLC.DefaultUni `Includes` PlutusRepr a) => PLiftable (PXor a)
+    ( PLiftable a
+    , PLC.DefaultUni `Includes` PlutusRepr a
+    , KnownBuiltinType (UPLC.Term UPLC.DeBruijn UPLC.DefaultUni UPLC.DefaultFun ()) (PlutusRepr a)
+    ) =>
+    PLiftable (PXor a)
 
 -- | @since 1.10.0
 instance PSemigroup (PXor PBool) where