+++ flake.nix
@@ -59,7 +59,7 @@
             src = ./.;
             compiler-nix-name = "ghc984";
             # NOTE(bladyjoker): Follow https://github.com/input-output-hk/plutus/blob/master/cabal.project
-            index-state = "2025-09-22T06:47:49Z";
+            index-state = "2026-03-18T14:20:24Z";
             inputMap = {
               "https://chap.intersectmbo.org/" = CHaP;
             };